split geometry into submodules
